Silence two warnings in the vendored pico_shared drivers
Both showed up in the PSRAM builds and neither is ours, but they are noise in
front of warnings that are, and this build has already cost one bug that gcc had
reported and a filtered log threw away.

hstx.h declared video_output_core1_run() a second time, in a header that already
includes video_output.h where it properly lives, so -Wredundant-decls fired in
every TU that saw both. Straight delete; nothing loses the declaration.

SetupPsram.c printed a uint32_t with %x. Cast rather than %08lx: uint32_t is long
unsigned on this target but plain unsigned elsewhere, so only the cast is right
in both places.

Both are recorded in 3rdparty/pico_shared_drivers/CMakeLists.txt with the other
local patches, since a resync re-copies from pico_shared and would drop them.
That list previously named only the pico_audio_i2s target rename and had lost
track of nespad_read_ready(), which is added back here.

A clean adafruitdvisd build, standalone and _full, is now down to the four known
pre-existing warnings: implicit malloc in i_video.c, 'skip_top' may be used
uninitialized in v_video.c, the deprecated tuh_init in d_main.c, and r_segs.c's
own #warning.

+++ 3rdparty/pico_shared_drivers/drivers/pico_psram/SetupPsram.c
@@ -210,7 +210,9 @@ int32_t __no_inline_not_in_flash_func(SetupPsram)(int psramCS)
     if (readback != 0x12345678)
     {
         _psram_size = 0;
-        printf("PSRAM test failed, readback = 0x%08x\n", readback);
+        // Cast, not %08lx: uint32_t is long unsigned here but plain unsigned on
+        // other targets, so the cast is the portable way to match %x.
+        printf("PSRAM test failed, readback = 0x%08x\n", (unsigned)readback);
         return 0;
     }
     return _psram_size;
